refactor: header 태그를 TopBar로 컴포넌트 분리
